Simplify installation with `gke-mcp install gemini-cli` command (#19)
* Simplify installation with `gke-mcp install gemini-cli` command

This commit introduces a new `gke-mcp install gemini-cli` command to streamline the installation process. This command registers the `gke-mcp` server as a `gemini-cli` extension, making it easier for users to get started.

Added an initial sketch of the extension context file. More work is needed to turn it into something really useful.

* Fix url in `go install` command.

* * Fix wording in GEMINI.md
* Add instructions for local development.